project05/requirements.txt
PyQt5>=5.12.0
mutagen>=1.45.0
Pillow>=8.0.0
